add moment template banner example stories
